Пропуск циклов отправки
